+++ RELEASE_NOTES.md
@@ -5,8 +5,9 @@
 ## 0.9
 
 - This version moves StoreKit logic to StoreKitPlus. 
-- This version adds a new subscription screen style. 
+- This version adds a new subscription screen style.  
 - This version adds purchased logic to the subscription screen.
+- This version adds a new `PremiumToggle` and `DisclosureToggle`.
 - This version wraps subscription screen types within the screen.
 - This version lets you add alternate primary onboarding buttons.
 - This version lets you specify which welcome onboarding to present.